requirements.txt
flask
flask-cors
requests
python-dotenv
gunicorn
aiogram==3.4.1
sqlalchemy